delete unfinished test script
